Refactor into modules and add entry point
