+++ README.md
@@ -8,7 +8,7 @@ Install requirements and run the application:
 
 ```bash
 pip install -r requirements.txt
-python word_copywriter.py
+python main.py
 ```
 
 1. Select the source document containing key-value pairs (lines formatted as `Key: Value`). PDF files are supported, including scanned documents (requires `tesseract` and `poppler` installed).
