+++ main.py
@@ -1,0 +1,15 @@
+import sys
+from PyQt5 import QtWidgets
+
+from gui import MainWindow
+
+
+def main():
+    app = QtWidgets.QApplication(sys.argv)
+    window = MainWindow()
+    window.show()
+    sys.exit(app.exec_())
+
+
+if __name__ == "__main__":
+    main()
